GestionarViajes - by cc
- Gestionar : Metodo de gestionarViajes teniendo encuenta los viajes activos (aun sin realizar) le pregunta si desea cambiarlo o cancelarlo : : : Metodo para cambiarlo aun sin realizar : : :
- Usuario : agregarSaldo puede ser un double
- Actualizacion de Date por LocalDate y metodos

 - *DESPIDO DE UN EMPLEADO CAUSA  PROBLEMAS - REVISAR*

+++ src/gestorAplicacion/Usuario.java
@@ -31,7 +31,7 @@ public abstract class Usuario implements Serializable {
         return this.billetera;
     }
 
-    public int agregarSaldo(int dinero){
+    public double agregarSaldo(double dinero){
         if(dinero > 0){
             this.billetera += dinero;
         }else{
